til_settings: rework setting get/add for bare values
The core thing here is rather than turning a bare value into a
key as I was doing before - we just leave the bare value as a
bare value and its setting must be located positionally via
get_value_by_idx since there's no key.

Existing callers that used to get_key() positionally now
get_value_by_idx() positionally all the same, except it's the
value instead of the key.  This is mostly done for things like
the module or fb name at the front of a settings instance.

The impetus for this change is partially just
cosmetic/ergonomics, but it's also rather strange for what's
really a key-less value to be treated as a value-less key.  It
was also awkward to talk/reason about on the road to recursive
settings where bare values would be supported as a standalone
settings instance if properly escaped...

This also adds unescaping of keys, and adds a dependency on the
somewhat linux-specific open_memstream() which may need changing
in the future (see comments).