til_settings: don't check values on nested settings
There needs to be more flexibility in the value checking
enforcement.

This is just a quick blunt-hammer fix to not trip over nested
settings values which will be huge undeterministic messes vs.
what's likely just a set of simple presets in the spec.values[]

The individual leaf settings will still be checked if they
specify values.  So this change just stops even bothering to
check unless the setting is a leaf.

This area needs more work in general, see comments.

For instance right now we can't just pass in arbitrary floats for
settings which list float values, not if that arbitrary float
isn't a member of the list.  In some circumstances that's the
right thing to do, as in the module can only work with the
presets.

But most of the time, the module would be perfectly happy with e.g.
foo=.3379 with foo's spec.values[] = { .01, .1, .25, .75, 1} so
the check fails.  That's dumb, and interferes with the creative
process when you're just poking different numbers into the
settings to see what happens.

TODO