til_settings: recursive til_settings_as_arg()
Currently in rototiller the only (de)serialization format of
settings is the args strings, so this is a rather critical piece
to get in for recursive settings to really be usable.

This is a quick and dirty implementation utilizing
open_memstream() which despite being POSIX has spotty support (I
don't think MacOS has it for instance).  So that's probably
something to rip out in the future.

Nonetheless, this moves things forward and works fine on GNU.