modules/compose: settings-ize layers and texture
Now layers= is a settings instance. Each individual setting
within that layers instance is also a settings instance of its
own.

This enables specifying the modules used in the layers as well as
settings to be passed into those per-layer modules.

The escaping quickly becomes brutal if hand-constructing, but
programmatically at least it's workable.  Plus, you can let the
interactive setup ask you for all the layer settings then just
copy and paste the cli invocation printed @ startup (at least
with rototiller).

texture= is also now a settings instance, which means compose no
longer randomizes the texture settings on its own - it instead
uses the settings supplied.  A consequence of this is that
texture settings need to be actually populated if the texture is
used.

For rtv, which randomizes settings, it makes no difference and
rtv compose invocations w/textures will just end up randomizing
the texture through the normal setup randomizing machinery.

But for direct compose invocations for instance, there's now an
actual texture setup process - and if you just use --defaults,
the defaults will be applied which is different from before where
it would have always been randomized.

This area needs some work, like controlling how defaults are
applied perhaps in the actual settings syntax such that
randomizing can still be performed if desired instead of
"preferred" defaults.  That's a more general settings syntax
problem to investigate