modules/rtv: add basic log_channels= setting
This introduces a boolean style log_channels= setting for
enabling logging of channel settings on channel switch.

It might be nice to change this to accept stdout/stderr/fdnum as
the setting instead of always directing at stderr.

This also doesn't capture the seed state so it's not exactly
logging everything needed to reproduce wholly what is being
shown.  Some compositions depend more on rand than others, so
it matters at varying degrees.

It'd be nice for settings syntax to have some global syntax
supported where a seed can always be embedded to be loaded.
Introducing such things as global settings to the settings syntax
is a pending TODO item... right now the only way to load seed
state is at startup passed to main as --seed=.  That's not gonna
cut it long-term.

This is an easy big step in the right direction though.  Trying
to make sense of what's on-screen from the truncated captions is
impossible.  Even if the captions wrapped the settings, it would
be tricky to catch the settings without recording the output or
screenshotting.

This also immediately makes me wonder about the voting system for
rtv where we log settings of favorites... then roll those into
playlists.