modules/compose: pass n_cpus to layers/texture context creates
Existing code was passing 0 which turns into the number of
cores/threads.

That's fine when compose isn't running nested in an already
threaded render, but falls down in something like checkers
w/fill_module=compose since checkers is already threading.  But
when checkers creates its fill_module context, it's careful to
pass 1 for n_cpus to prevent that kind of thing.  With this
change that no longer falls apart.