til_setup: refcount til_setup_t
The whole point of til_setup_t is to represent the baked, most
conveniently usable form of a setup derived from one or more
settings instances.  Things generally go from the serialization
format "settings string" to til_settings_t eventually culminating
in a til_setup_t.

So the process of making a til_setup_t is rather tedious and kind
of costly.  Once into a til_setup_t it's desirable to just hang
on to this form and reuse it if possible.  The way a til_setup_t
baked setup is put to use is in a read-only fashion where it
basically just informs behavior, so it makes a lot of sense to
enable refcounting the thing and letting whatever can make use of
it bump the refcount and hold onto the pointer, accessing the
contents whenever it needs to answer a question about that
particular setup.

The immediate impetus for this is actually rtv's snow_module
setup.  In rtv every channel switch may recreate the context, if
the context has expired.  In the case of the snow module, the
context always expires, and we definitely want to discard the
context while playing the next channel.  But when the snow
resumes, in order to recreate the context as configured, we need
the same setup again.  It just becomes clear that what's needed
is a way to pin the snow_module's setup for this reuse to be
safe.

There's also plenty of other modules that have been piecemeal
copying settings into their context, when what would really make
more sense is to just ref it and stow the pointer, then unref on
their context destroy.  They can just access the setup via the
pointer as needed, instead of having to duplicate the setup in
their context.  Indeed, some module contexts even embed the
entire setup just to copy its contents over by value.  In
simple/small scenarios that's fine, and I'm sure in those
particular cases it's perfectly safe to do.  It just seems
unnecessary altogether.

Another small change made is supporting NULL free_func, which
will default to libc's free().  Most til_setup_new() call sites
are passing free() with an annoying cast, those can be changed to
NULL.