til: add setup assert in til_module_create_context()
If the module has a .setup() method, it should be able to just
blindly expect a valid baked setup to come in @ context create.

Let's enforce that here, and let the module code providing setup
hooks just go forth and access the setup assuming it's at least
present and as valid as their setup method produces.