modules/*: remove use of static default setups
There was a time when it made sense for context creates needing
setups but not receiving them to still be functional with some
sane defaults.

But with recursive settings, we really shouldn't ever have
orphaned nested module uses unreachable by a proper setup.

So let's just get rid of this fallback, and exclusively rely on
the baked setups provided by the .setup() methods.  They still
have preferred defaults, and the proper setup production
machinery is what should be responsible for applying those
at runtime where they may also be overridden or otherwise
influenced.