modules/*: stop storing setup by value in contexts
With setup refcounting and a reference bound to the context, we
should just dereference the single instance.  The way setups are
used it just as a read-only thing to affect context behavior...

Note I've left the module-type-specific setup pointer despite it
duplicating the setup pointer in the module_context.  This is
just a convenience thing so the accessors don't have to cast the
general til_setup_t* to my_module_setup_t* everywhere.