til_fb: clip get_pixel_clipped() to fragment
til_fb_fragment_get_pixel_clipped() was clipping to the frame
bounds, which are actually just a logical dimension for placing
and scaling fragments within a frame.  drizzle in particular is
leaning on this clipped-get to prevent accessing outside of the
fragment's backing buffer.  Let's just clip it to the fragment
bounds instead, see comment added for more information.