til_settings: drop settings->label from generated label
Since these are ultimately intended for use in path construction,
it's redundant to include the settings->label in the generated
label.  Instead what's really useful is just the subscript part:

 /module/compose/layers/layers[0]/drizzle/viscosity

 Becomes:
 /module/compose/layers/[0]/drizzle/viscosity

which is far better.  It may seem silly to have both the
positional subscript *and* the module name, as in why not just
have:

 /module/compose/layers/drizzle/viscosity

But there's a need ot handle potential collisions like so:

 /module/compose/layers/[0]/drizzle/viscosity
 /module/compose/layers/[1]/stars
 /module/compose/layers/[2]/drizzle/viscosity

So then maybe you think; ok, why have the module name? just use
the positional subscript since that alone prevents the
collisions.  Result:

 /module/compose/layers/[0]/viscosity
 ...
 /module/compose/layers/[2]/viscosity

Well, now we've lost useful context.  The viscosity setting
recurs on multiple modules, and we don't know just at a glance
what we're working with anymore.

Hence, why there's both.  The module name in the path makes
things substantially more self-explanatory.  These paths will
likely be what you're looking at as the labels of tracks in a
multitrack sequencer like GNU Rocket.  So this decision is likely
affecting the UX at that level in the fullness of time.