modules/checkers: desc the fill_module's name setting
The bare-value value_as_nested_settings.entries[0] setting which
serves as the name for module lookup is in a sort of no-mans land
between checkers and the underlying fill_module's setup.

So we have to do this little bit of rigamarole in checkers, being
the entity wiring up the nested module.  The fill_module's
setup_func won't be doing anything to describe the name's
setting as it's only interested in its own settings.

There will likely be some helpers made later to streamline this
process of composing module/settings hierarchies.