til_settings: add some settings and desc path printers
These print to a stdio FILE*, which for now is harmless since
setup_interactively() will be the only consumer (... for now).

If one needed the path in a buffer, they could use
open_memstream() and pass that FILE* to the printers.  And since
commit 40feb61 there's already a dependency on the function, but
it's icky to dig deeper into that portability trap.

Although it seemed like rototiller actually compiled on MacOS at
Gene's house, so maybe open_memstream() isn't as problematic as
it once was.

Anywhow, this commit only adds the path printing helpers, nothing
is using them yet.