modules/rtv: don't use n_cpus=0 for the  context creates
This is harmless as long as rtv stays hermetic.

But if rtv gets used in a composite scenario in the future by
either removing the hermetic flag, or allowing forced overrides,
n_cpus=0 will cause the nested module contexts to become threaded
on SMP machines.

That's problematic if the outer module context is already a
threaded render.  What's appropriate here is to just propagate
the n_cpus down so if an upper layer has already gone threaded,
it will be sending down n_cpus=1 to serialize the nested
instances.

In practice, as-is, this change basically changes nothing, but
prepares for a potential future where rtv participates in
threaded compositions.

Through a lens of "rtv just rejiggers scenes and there settings
on a timer from a settings-specified subset of modules and
settings" it's arguably useful as just another module.  Sometimes
you want something to change itself up periodically in say a
compose layer.

So preparing for this possibility isn't really all that
far-fetched/hypothetical.