modules/rtv: desc the snow_module name setting
Like modules/checkers required for fill_module, we need to do the
same for for rtv::snow_module.

There's more work to do on rtv::channels, but that's still
unsettled stuff in terms of settings syntax since rtv randomizes
settings.  It's desirable to have the rtv settings able to
specify which settings to hold constant at a specific value
per-channel, leaving everything else for randomizing on channel
switch.

But there's no syntax for that kind of stuff currently, and it
seems like there's a need to communicate during the setup_func
dance when we're in a "settings optional because we'll fill them
in automatically at time of use later" to the front-end.

It's not strictly a front-end issue though - because the back-end
setup_func actually controls the forward progress.  From the
current setup_func's perspective, everything's important to it
and must be fulfilled.  And we certainly want the setup_func to
continue informing the setup process.

So it's more like the channel settings being populated via rtv
still need to all get populated, rtv just needs a way to add an
attribute to mark which settings are static vs. which should get
randomized on every use.

Perhaps there should just be a special value syntax reserved for
saying "random value" and the front-end can apply that, but then
a til_module_randomize_setup() could detect that too in a
per-setting flag the front-end set.  That way the value gets
re-randomized, while the ones without that value set get left
alone.

Yes, I know this isn't the appropriate place for such commentary.
But nobody is reading these things anyways on my toy side project.