modules/voronoi: remove dirty= setting and variant
This variant is kind of a broken hack, and its brokenness becomes
more apparent in a threaded voronoi world.

So just drop it for now.

I am interested in more voronoi variants, but they can't
compromise correctness/introduce instabilities or significantly
interfere with performance improvements like threaded rendering.

The dithered-ish look of dirty=on was an interesting variant
though... bummer.