til,modules/rtv: supply settings to til_module_setup_randomize()
Rather than creating an orphaned settings instance private to
til_module_setup_randomize(), the function now requires the
settings instance be provided.

The one remaining caller of this function is modules::rtv.  Now
that rtv is responsible for creating the settings beforehand, and
the settings may be created with a path prefix, rtv gets its
til_module_context_t->setup.path prefixed for all the channel
settings.

Another improvement is now the channel settings instance gets
created from the module name as the settings string.  So while
it's not yet possible to sparsely specify settings with others
being randomized, at least now when log_channels=on is in effect,
the printed args include the top-level channel module.

Having proper complete paths for the rtv channel modules is
especially visible in --print-paths output FYI.

An interesting test for exercising all this is:

```
$ src/rototiller --module=rtv,duration=0,context_duration=0,snow_module=none,channels=all,log_channels=on --print-pipes --defaults --go 2>/tmp/channels

in another terminal:
$ tail -F /tmp/channels
```

watch the chaos unfold