til: introduce "ref" built-in module
This opens up the possibility of referencing contexts on-stream
by-path anywhere a module may be used:

ref,path=/path/to/existing/context

Currently the path lookup is performed @ render time, and a
reference is taken on the context found there on the first time
it's located.  That reference is then held in the ref's context
until its context is destroyed.

There are more details to flesh out, and what probably needs to
happen is some til_stream API for maintaining the reference in
the event that the context at that path gets replaced.  Due to
refcounting, the referenced context will persist even if the
stream-resident instance goes away, but there should probably be
a way for that replacement to invalidate the existing references
so they replace theirs with a fresh lookup at the path...

Either way this is a good start.  With this addition you can
effectively implement scene transitions in rkt, via the mixer
module, something like (sans proper escaping):

mixer,a_module=ref\,path=/foo/previous/ctxt,b_module=ref\,path=/foo/next/ctxt