til: register contexts on-stream @ create time
Module contexts are now discoverable via the stream a la
til_stream_find_module_contexts().

The current architecture has the stream adding a reference to all
the contexts registered.  So even if they get unreferenced by
their creators, they will linger on-stream.

There's a gc mechanism til_stream_gc_module_contexts() which can
be used to trigger a cleanup of contexts _only_ referenced by the
stream.

It's unclear as of yet if this is the way to go long-term, but it
lets things work for now and allows some iterating and
experimentation to see where to go next.