modules/rkt: handle stream taps @ context create
It's problematic getting this stuff online at render time,
because the modules end up rendering with uninitialized tap
values in that first frame.

With this change modules can get their taps on-stream at context
create, and we'll notice and do the initial rocket update just
before returning the rkt context to prime everyone on-stream.