build: adding -D_POSIX to CFLAGS for mingw's rand_r
I'm not sure if this will stay long-term, but it seems not be
disrupting linux builds and with open_memstream all gone now
mingw is succeeding with this present.

It would really make sense to simply replace all rand_r() usage
with a libtil rand_r equivalent.  That way it can guaranteen
consistent behavior across platforms...