til_module_context: s/ticks/last_ticks/ in context
And just maintain it as the last ticks value after rendering with
the context...

A couple modules were already doing this manually in an ad-hoc
fashion, just make it a general thing.

Updated those modules to reflect the new situation

Especially in a rkt world with modules::mixer doing fades, it
becomes common to render the same context twice in the same frame
for the blending.  We need to prevent accelerated animations in
such situations.  For now let's just rely on ticks in a delta-T
fashion to prevent animating the context when ticks is the same.

modules::stars in particular needs this fixed up, upcoming commit