modules/rkt: end the stream on scene 99999
This conditionally will end the stream on scene 99999 if
connect=off (playback mode)

When connect=on it'll just make it show an "EXIT SCENE"
diagnostic instead of the "NO SCENE" message.

Now you just stick 99999 in the rkt:scene track to end the show.

It's assumed 99999 scenes will never be needed...