modules/rkt: drop vestigial TODO comment about handles
This was written when module names were going to have an
@/path/to/context "handle" syntax.  But instead I went the "ref"
builtin module route, with path=/path/to/context as a setting.

While it's more verbose in the settings, it "just works"
everywhere that can take a module+settings because the ref
builtin is just another module like any other.

So this TODO is referring to something that won't happen in a
"ref" builtin world.