modules/moire: lose unnecessary branch in inner loop
I had benchmarked this change and it showed no difference at all
on my 2c/4c i7 X230.

But having just tried it on an RPi4B where it moved the test case
from 54FPS to 60FPS, a +10% improvement, it's worth the
readability loss.

It's interesting how Intel's cleverness discourages optimizing in
ways that benefit probably *all* the competition... even when the
optimization is such a minor change in terms of effort.