til: use 16 * n_cpus in til_fragmenter_slice_per_cpu()
Slight improvement of CPU utilization for fragmenters using this
strategy...

I noticed tile64 would give better FPS in some scenarios where it
seemed obvious slice_per_cpu() was the appropriate option.  And
that turned out to just be by virtue of being able to give idle
threads something to do while busy ones finished what was on
their plate.

So just make the slices a bit more granular than n_cpus... this
may have to be revisited in the future to find the sweet spot,
and may need to be more sophisticated than just multiplying by a
constant factor.