til_settings: add way to change the settings label
In rkt_scener it's desirable to fully construct a new settings
instance *before* wiring it into the parent rkt/scenes settings.

But the way these positional settings labels get constructed
depends on the parent's settings entries.  So it becomes a
chicken-egg problem to require the proper label @ creation before
its containing setting is even added in the parent's entries.

With this change a temporary WIP label can be used while
constructing the settings instance, then amend it once we know
the settings instance is complete and ready to join the party.