til: make til_modlue_setup() nested usage aware
When setting up a nested module, the HERMETIC modules shouldn't
be considered as eligible nor should one be the preffered option
(rtv).

This uses the presence of a parent on the settings as a heuristic
for if it's a nested scenario (no parent == root).  When nested
the default changes to "compose" from "rtv", and HERMETIC modules
are omitted from the listed values.