setup: return failed desc _path_ from setup_interactively()
Returning the failed desc was just a lazy half-assed thing that
was sort of the best option in the simpler, pre-paths world.

But now that everything has paths thanks to recursive settings,
let's just return the path to the failed setting's desc.

This conveniently gets rid of a UAF bug when setup returned the
setting->desc as the failed desc, and main would print the desc
*after* freeing all the settings in its final moments.

But the best part is now more of the errors parsing settings
should be accompanied by an illuminatingly relevant setting path.

Previously you'd at best get a bare key from the desc, but often
no failed desc was returned at all and you saw no guidance at
all.  But with the recent improvements to the setup error
handling I think those cases should be few if not entirely
eliminated.