modules/rkt: basic support for Rocket reconnects
This changes things so rkt won't exit with an error @ startup if
RocketEditor isn't already listening.  It also tolerates
RocketEditor going away, and will show a "OFFLINE" overlay status
text should that happen w/connect=on.

Some status text has also been added to the "EXIT SCENE" 99999
scene for both the RocketEditor connection and the scener
enabled/disabled status.  No indicator yet for if scener has a
connection though, only if it's listening or not via listen=on.