modules/rkt: fix recv loop to continue on non-NL
After putting the recv() in a for(;;) to not have to render a
frame per byte received, I completely dropped the ball on moving
the return and adding the continue to actually finish the change.

This makes creating new scenes via pasting long settings strings
far less laggy.  A future improvement would be to not recv() a
byte at a time, but this really isn't a perf-sensitive thing.