modules/shapes: make threaded
This is just a first stab at threading shapes... whenever shapes
finds itself in a scene it easily becomes a significant
bottleneck, threading is a trivial path to improving that
somewhat.

While at it I also got rid of the need for _checked() variants
which also helps a bit.

But nothing here is proper optimization of the routines.. there's
too much math happening per pixel in a naive fashion.