til: measure til_module_render() durations
This stows the duration (in ticks (which are ms for now)) of a
given module context's most recent, as well as tracking the max,
in til_module_context_t.

For now it's also added to --print-module-contexts output, but
this is still rather primitive and nearly inscrutible in
practice...  that output is a flickery and unsorted mess during
playback.  But this is just a start.

Ticks may need to move up to microseconds if it'll also be the
units for measuring timings.  Right now things are slow enough
that the low-hanging fruit stand out as multiple if not dozens of
milliseconds so it's still useful for now.