modules/spiro: skip color math+packing if textured
The color stuff as-is isn't cheap and doesn't even get used if
there's a texture present, so don't bother with it at all.

This is especially significant since this module isn't threaded,
so it ties up all the cores leaving most of them idle when part
of a composition.

Also since spiro doesn't clamp its coordinates to the fragment
dimensions, only considering the frame dimensions, it must
continue using the "checked" put_pixel variant...