til: introduce "noop" builtin
When profiling compositions it's convenient to suppress some
modules in the settings to see how their omission affects the
FPS.

There's already a "blank" builtin which clears the fragment, but
there's no way to turn the cost to ~zero.  This "noop" builtin
does just that, absolutely nothing - it won't even trigger the
implicit "til_fb_fragment_t.cleared = 1" update so whatever
rendering follows will still find an uncleared fragment if that's
what came into the noop.