til_fb: fast-path til_fb_fragment_clear() a bit
What this really should have is a general memset32()
implementation, but for now it's a small improvement at least for
clearing since that's known to be a 0-pixel case where memset()
is applicable.

memset() isn't usable for actual RGBA pixels since it operates
strictly on bytes.  There's some hacks for using memmove() to get
closer where you store the first four bytes, then memove the
start of the buffer to the remaining, doubling the size of the
move each time for log(n) calls to memmove.  Maybe that makes
sense as a stop-gap for the general case, it needs
implementing+benchmarking first.

This commit is dead simple though and an obvious little FPS gain
for clearing.